requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
distlib==0.3.4
docopt==0.6.2
filelock==3.6.0
idna==3.3
platformdirs==2.5.2
requests==2.27.1
six==1.16.0
urllib3==1.26.9
yarg==0.1.9
py-cord==2.0.0b7
Pillow==9.1.0
schedule==1.1.0
python-decouple==3.6
numpy==1.22.3
py-cpuinfo==8.0.0
psutil==5.9.0
discord-webhook
pytz
bs4
